requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
pydantic==2.10.2
pytest==8.3.3
pytest-asyncio==0.24.0
